Move to prod code testfixtures
