+++ acceptance-tests/build.gradle
@@ -11,6 +11,9 @@
  * specific language governing permissions and limitations under the License.
  */
 
+apply plugin: 'java-library'
+apply plugin: 'java-test-fixtures'
+
 dependencies {
 
   testRuntimeOnly 'javax.activation:activation'
@@ -27,8 +30,6 @@ dependencies {
 
   testImplementation 'org.apache.logging.log4j:log4j-api'
   testImplementation 'org.assertj:assertj-core'
-  testImplementation 'org.awaitility:awaitility'
-  testImplementation 'org.apache.tuweni:tuweni-net'
   testImplementation(testFixtures(project(":keystorage:hashicorp")))
 }
