+++ keystorage/hashicorp/src/testFixtures/java/tech/pegasys/signers/hashicorp/dsl/certificates/CertificateHelpers.java
@@ -10,7 +10,7 @@
  * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
  * specific language governing permissions and limitations under the License.
  */
-package tech.pegasys.signers.dsl.certificates;
+package tech.pegasys.signers.hashicorp.dsl.certificates;
 
 import java.io.FileOutputStream;
 import java.io.IOException;
