+++ keystorage/hashicorp/src/testFixtures/java/tech/pegasys/signers/hashicorp/dsl/hashicorp/HashicorpVaultDockerCertificate.java
@@ -10,11 +10,11 @@
  * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
  * specific language governing permissions and limitations under the License.
  */
-package tech.pegasys.signers.dsl.hashicorp;
+package tech.pegasys.signers.hashicorp.dsl.hashicorp;
 
 import static java.nio.file.Files.createTempDirectory;
 
-import tech.pegasys.signers.dsl.certificates.SelfSignedCertificate;
+import tech.pegasys.signers.hashicorp.dsl.certificates.SelfSignedCertificate;
 
 import java.io.IOException;
 import java.nio.file.Path;
